+++ packages/data-types/src/lib/interfaces/heart-event.interface.ts
@@ -1,0 +1,7 @@
+import { HeartEventTriggerData } from "@ci/data-types";
+
+export interface HeartEvent {
+    npc: string,
+    heartLevel: number;
+    trigger: Omit<HeartEventTriggerData, 'npc' | 'heartLevel'>[]
+}