+++ packages/guide/src/app/npcs/components/heart-events/heart-events.component.ts
@@ -1,0 +1,12 @@
+import { Component, Input } from '@angular/core';
+import { HeartEvent } from "@ci/data-types";
+
+@Component({
+    selector: 'app-heart-events',
+    templateUrl: './heart-events.component.html',
+})
+export class HeartEventsComponent {
+
+    @Input({required: true}) heartEvents!: HeartEvent[];
+
+}